refactor(auth): move the single grant-revocation body to the shared tier
revoke_grant_and_sweep_tokens moves from auth/services/oauth_grant_service
to shared/oauth_grant_revocation (the tier where the session-taking
audit/event helpers already live), re-exported from its old home so every
auth caller is unchanged. Needed so the admin OAuth-client hard delete can
run THE single revocation body without violating the admin-must-not-import-
auth boundary or the no-sqlalchemy-in-admin-services rule.
